requirements.txt
# pip install -r requirements.txt
# pip install --upgrade django

psycopg2-binary
django-on-heroku
asgiref==3.7.2
Django~=5.0.2
sqlparse==0.4.4
djangorestframework~=3.14.0
gunicorn
commonmark
requests
django-background-tasks
drf_yasg